refactor: 글목록 컴포넌트 리팩토링, 재사용
