+++ views/components/post-list/styles.ts
@@ -21,4 +21,5 @@ export const Pgbox = styled.div`
   display: flex;
   flex-direction: column;
   align-items: center;
+  width: 100%auto;
 `;
