--- views/index/diarybox/styles.ts
@@ -1,25 +1,0 @@
-import styled from 'styled-components';
-
-export const DiaryBox = styled.div`
-  height: 30px;
-  width: 100%;
-  margin-bottom: 10px;
-  cursor: pointer;
-  background: #f3e1e1;
-`;
-
-export const contentbox = styled.div`
-  float: left;
-  margin: 5px;
-`;
-
-export const forleft = styled.div`
-  float: left;
-  margin: 5px;
-`;
-
-export const datebox = styled.div`
-  float: right;
-  margin-right: 10px;
-  margin: 5px;
-`;
