--- views/index/emptybox/styles.ts
@@ -1,19 +1,0 @@
-import styled from 'styled-components';
-
-export const EmptyBox = styled.div`
-  width: 100%;
-  height: 100px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-top: 30px;
-`;
-
-export const EmptyBoxTitle = styled.div`
-  font-weight: 700;
-  margin-bottom: 10px;
-`;
-
-export const EmptyBoxText = styled.div`
-  margin-bottom: 10px;
-`;
