+++ views/index/index.tsx
@@ -2,14 +2,11 @@ import Layout from 'views/components/layout/index';
 import { NextPage } from 'next';
 import * as S from './styles';
 import Sidebar from 'views/components/sidebar';
-import Diarybox from './diarybox';
-import React from 'react';
-import ReactPaginate from 'react-paginate';
 import postAPI from 'common/api/postAPI';
-import { usePagedPosts } from './hooks';
-import Emptybox from './emptybox';
 import { PagedPosts, Post } from 'share/interfaces/post';
 import MoodCalendar from './mood-calendar';
+import PostList from 'views/components/post-list';
+import EmptyBox from 'views/components/post-list/empty-box';
 
 interface IndexPageProps {
   initialPosts: Post[];
@@ -17,8 +14,6 @@ interface IndexPageProps {
 }
 
 const IndexPage: NextPage<IndexPageProps> = ({ initialPosts, total }) => {
-  const { pageCount, changePage, pagedPosts } = usePagedPosts({ initialPosts, total });
-
   return (
     <Layout>
       <Sidebar />
@@ -28,30 +23,8 @@ const IndexPage: NextPage<IndexPageProps> = ({ initialPosts, total }) => {
         </S.CalendarContainer>
 
         <S.DiaryListContainer>
-          <S.Diaryinfo>나의 일기들</S.Diaryinfo>
-          <S.DiaryBoxContainer>
-            {pagedPosts.length ? (
-              pagedPosts.map((post) => {
-                return <Diarybox key={post.id} post={post} />;
-              })
-            ) : (
-              <Emptybox />
-            )}
-          </S.DiaryBoxContainer>
-          <S.Pgbox>
-            {pageCount > 0 && (
-              <ReactPaginate
-                pageRangeDisplayed={2}
-                marginPagesDisplayed={3}
-                previousLabel={'이전'}
-                nextLabel={'다음'}
-                pageCount={pageCount}
-                onPageChange={changePage}
-                containerClassName={'pagebtn'}
-                activeClassName={'page_active_btn'}
-              />
-            )}
-          </S.Pgbox>
+          <S.Diaryinfo>목록 보기</S.Diaryinfo>
+          {initialPosts ? <PostList initialPosts={initialPosts} total={total} /> : <EmptyBox />}
         </S.DiaryListContainer>
       </S.Mainpage>
     </Layout>
